pyproject.toml
[tool.poetry]
name = "syncer-api"
version = "0.1.0"
description = ""
authors = ["Ajeem Ahmad <ahmedazim7804@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.13"
loguru = "^0.7.3"
sqlmodel = "^0.0.24"
grpcio = {extras = ["asyncio"], version = "^1.73.0"}
grpcio-tools = "^1.73.0"
grpc-interceptor = "^0.15.4"
python-dotenv = "^1.1.0"
dependency-injector = "^4.48.1"
uvicorn = "^0.34.3"
python-multipart = "^0.0.20"
pyjwt = "^2.10.1"
ruff = "^0.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
